Cleaning up the example of current strategy
You have the right idea with the example you show (the current way of having to setup the 3 action creators), but I think you can simplify the API call example a bit. This is more likely what someone would write. It's a little shorter, but still not as clean as your solution :)